Moved chat/commands
